Fixes for building on Android.
Fixed search paths for some Find*.cmake files.
Added FindOpenGLES2.cmake. find_package(OpenGL) won't find OpenGL ES.
More robust detection for when spinlocks are available in pthread.
Workaround for lack of qsort_r using thread-local storage with qsort.
Warning fixes for 32-bit systems.

+++ modules/Core/src/Streams/Stream.c
@@ -143,7 +143,7 @@ uint64_t dsStream_skip(dsStream* stream, uint64_t size)
 			if (readSize > sizeof(buffer))
 				readSize = sizeof(buffer);
 
-			uint64_t thisReadSize = dsStream_read(stream, buffer, readSize);
+			uint64_t thisReadSize = dsStream_read(stream, buffer, (size_t)readSize);
 			totalSize += thisReadSize;
 			if (thisReadSize != readSize)
 				return totalSize;
